Replace CreateCustomerCommand with ProviderBillingService and move to Commercial
